iot/l3/requirements.txt
anyio==3.6.2
appdirs==1.4.4
application-utility==1.3.2
asgiref==3.5.2
attrs==22.1.0
autocommand==2.2.1
btrfsutil==6.0
CacheControl==0.12.11
certifi==2022.9.24
cffi==1.15.1
chardet==5.0.0
click==8.1.3
colorama==0.4.6
commonmark==0.9.1
cryptography==38.0.1
cupshelpers==1.0
Cython==0.29.32
distlib==0.3.6
distro==1.7.0
docopt==0.6.2
docutils==0.19
dotenv-python==0.0.1
fastapi==0.86.0
fastapi-mqtt==1.0.5
future==0.18.2
gmqtt==0.6.11
greenlet==2.0.1
h11==0.14.0
idna==3.4
inflect==6.0.2
jaraco.context==4.1.2
jaraco.functools==3.5.2
jaraco.text==3.10.0
keyutils==0.6
lit==14.0.6.dev0
manjaro-sdk==0.1
more-itertools==9.0.0
msgpack==1.0.4
netsnmp-python==1.0a1
npyscreen==4.10.5
nspektr==0.4.0
ordered-set==4.1.0
packaging==21.3
pacman-mirrors==4.23.2
paho-mqtt==1.6.1
pep517==0.13.0
Pillow==9.3.0
platformdirs==2.5.2
ply==3.11
psutil==5.9.3
pycairo==1.21.0
pycparser==2.21
pycups==2.0.1
pycurl==7.45.1
pydantic==1.10.2
Pygments==2.13.0
PyGObject==3.42.2
pynvim==0.4.3
pyOpenSSL==22.1.0
pyparsing==3.0.9
PyQt5==5.15.7
PyQt5-sip==12.11.0
PySide6==6.4.0.1
pysmbc==1.0.23
python-dotenv==0.21.0
PyYAML==6.0
reportlab==3.6.11
requests==2.28.1
resolvelib==0.8.1
rich==12.6.0
shiboken6==6.4.0.1
six==1.16.0
sniffio==1.3.0
starlette==0.20.4
systemd-python==235
team==1.0
tenacity==8.1.0
tomli==2.0.1
trove-classifiers==2022.10.19
typing_extensions==4.4.0
udiskie==2.4.2
urllib3==1.26.12
uvicorn==0.17.6
validate-pyproject==0.10.1
webencodings==0.5.1
